Partially added player UI and added 'Room Mechanic' event to Master Room
If your room has an event/mechanic in it that would be 'press X to do the mechanic', use Event Room Mechanic and set the RoomHasMechanic to true and get me to set MechanicButtonStyle to an appropriate button visual. Also made some fixes to RedandBlueRoom - Callum you need to get rid of the 'entrance boxes' actors in the room before running the event Room Complete - they stop new rooms from spawning.